backend/requirements.txt
flask
python-dotenv
flask-restful
flask_sqlalchemy
flask_migrate
flask_jwt_extended
flask_mail 